+++ pilz_store_positions/test/unittests/test_data/missing_type_file.py
@@ -1,0 +1,38 @@
+# Copyright (c) 2020 Pilz GmbH & Co. KG
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Lesser General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+start_pose_test = PoseStamped(
+    header=Header(
+        seq=0,
+        stamp=Time(
+            secs=12345,
+            nsecs=6789
+        ),
+        frame_id='world'
+    ),
+    pose=Pose(
+        position=Point(
+            x=7,
+            y=0.0,
+            z=0.0
+        ),
+        orientation=Quaternion(
+            x=0.0,
+            y=0.0,
+            z=0.0,
+            w=-1.0
+        )
+    )
+)